+++ src/main/kotlin/dev/turingcomplete/intellijbytecodeplugin/view/_internal/_structure/_class/FramesPanel.kt
@@ -111,7 +111,7 @@ class FramesPanel(initialTypeNameRenderMode: TypeUtils.TypeNameRenderMode, metho
           add(ViewValueAction())
         }
       })
-      TableSpeedSearch(this)
+      TableSpeedSearch.installOn(this)
     }
   }